+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 - Stricter handling of invalid keys in Traversables passed to PSR-16
 - Allow integers to be used as keys for PSR-16
+- Treat a TTL of 0 as expired for PSR-16
 
 
 ## [1.4.1] - 2017-01-03